refactor(test): remove cross-step workflow history
