+++ packages/test/src/engine/index.ts
@@ -3,5 +3,4 @@ export * from './document-runtime';
 export * from './registry';
 export * from './run-workflow-document';
 export * from './types';
-export * from './history';
 export * from './project-runtime';
