+++ packages/test/tests/e2e/fixtures/test-project/flows/first.yaml
@@ -5,8 +5,6 @@ cases:
           value: first:one
       - test.record:
           value: first:two
-      - test.expect-history:
-          count: 2
 
   - name: second case
     steps:
