+++ packages/test/tests/e2e/fixtures/test-project/flows/nested/second.yml
@@ -5,5 +5,3 @@ cases:
           value: nested:one
       - test.record:
           value: nested:two
-      - test.expect-history:
-          count: 2
